+++ src/graphics/view_to_projection.c
@@ -42,7 +42,7 @@ void		view_to_projection(t_polygon *p, int count, t_mat4d proj_mat,
 	i = -1;
 	while (++i < count)
 	{
-		if (p[i].tex_id == -1)
+		if (p[i].tex_id == (Uint32)-1)
 			continue ;
 		p[i].v01 = mat4vec4_multiply(proj_mat, p[i].v01);
 		p[i].v12 = mat4vec4_multiply(proj_mat, p[i].v12);
